changed: code organization
